Move DraftStateMixin to draft_state submodule
